requirements.txt
typer==0.9.0
rich==13.5.2
zeroc-ice==3.6.5
omero-py==5.13.1
omero-upload==0.4.0
ezomero==2.1.0
